test(core): split browserslist build tests
